requirements.txt
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.0
ecdsa==0.14.1
fastapi==0.60.1
gunicorn==20.0.4
h11==0.9.0
idna==2.8
joblib==0.14.1
loguru==0.4.1
numpy==1.18.1
pandas==1.1.0
passlib==1.7.2
pyasn1==0.4.8
pycparser==2.20
pydantic==1.6.1
python-dateutil==2.8.1
python-jose==3.2.0
python-multipart==0.0.5
pytz==2020.1
requests==2.22.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn-pandas==2.0.0
starlette==0.13.6
threadpoolctl==2.1.0
urllib3==1.25.10
uvicorn==0.11.2
websockets==8.1
win32-setctime==1.0.1
